Fix Scripts
