+++ VideogameProject/Unity_FA/Assets/Scripts/CardManager.cs
@@ -21,33 +21,21 @@ TO-DO
 public class CardManager : MonoBehaviour
 {
     [SerializeField] GameObject CardPrefab;
-    [SerializeField] GameObject PuButtonPrefab;
-    [SerializeField] GameObject PUPrefab;
-
+    
     [SerializeField] Transform parentPrefab;
-    [SerializeField] Transform PUParentPrefab;
-
-    [SerializeField] Transform PUSlot;
-
+    
     [SerializeField] public List<GameObject> Cartas_mano = new List<GameObject>();
     [SerializeField] public List<GameObject> Panels = new List<GameObject>();
     
 
-    [SerializeField] List<GameObject> pu_Hand;
-    [SerializeField] List<int> pu_Pile;
-    [SerializeField] List<int> pu_Discarded;
-
-    [SerializeField] GameObject AddPUBtn;
-
-    int maxPuPile = 28;
-
+   
     [SerializeField] TMP_Text turnText;
     [SerializeField] TMP_Text energyText;
     [SerializeField]  GameObject energySlider;
 
 
 
-    public bool puAdded;
+   
 
     private bool changeButtonPressed = false;
     public bool PlayerTurn = false;
@@ -68,9 +56,9 @@ public class CardManager : MonoBehaviour
     public int energy;
     public int max_energy;
     [SerializeField]  Cards allCards;
-    [SerializeField]  PUs puCards;
+    
     public string APIData;
-    public string pu_Cards_Data;
+    //public string pu_Cards_Data;
 
     [SerializeField] public bool Dataready = false;
     [SerializeField] public bool PU_Dataready = false;
@@ -78,35 +66,7 @@ public class CardManager : MonoBehaviour
     [SerializeField] bool PowerUp_created = false;
     [SerializeField] bool pu_saved = false;
 
-
-    /*
-    PlayerTurn()
-    atributos:
-        GameObject activeCard1
-            -attack
-            -HP (slider o text)
-        GameObject activeCard2
-            -attack
-            -HP
-        GameObject selectedCard
-
-        methods:
-        attack() ->
-        change() -> 
-        end() -> end turn 
-
-
-        FLOW
-        1. Select an ally card
-        2. Choose to
-            a) attack
-            b) change
-        3. Select an enemy/ally card
-        4. Check if there are more options??
-            power ups? attack with the other card?
-        5. End Turn 
-
-    */
+    
 
     void Start()
     {
@@ -129,12 +89,39 @@ public class CardManager : MonoBehaviour
             GameStarted = true;
         }
     }
+
+    void InitGame()
+    {
+        
+        StartCoroutine(Create_Board());
+       
+    }
     //template of all  the cards in the board 
+    IEnumerator Create_Board()
+    {
+        int i = 0;
+        
+        allCards=JsonUtility.FromJson<Cards>(APIData);
+        
+        Debug.Log(allCards.cards);
+        
+        foreach (Atributos atributosCarta in allCards.cards)
+        {
+            Card_base(Cartas_mano, i, atributosCarta);
+            yield return new WaitForSeconds(0.1f);
+            i++;
+        }
+        
+        GetComponent<PU_controller>().Init_PU_ID();
+        
+        GetComponent<PU_controller>().PU_button();
+    
+        PlayerTurn = true;
+    }
 
     public void Card_base(List<GameObject> Lista, int i, Atributos atributosCarta)
     {
         GameObject Card = Instantiate(CardPrefab, Panels[i].transform.position, Quaternion.identity, parentPrefab);
-       // Card.AddComponent<Atributos>().SetAtributos(atributosCarta);
         Card.GetComponent<CardScript>().Init(atributosCarta);
         Card.GetComponent<CardScript>().selfCard=Card;
         Card.name = "Card" + Lista.Count;
@@ -143,266 +130,7 @@ public class CardManager : MonoBehaviour
         Button button = Card.GetComponent<Button>();
         button.onClick.AddListener(() => registerCard(Card));
         
-        
     }
-
-    //template of all the power ups in the board
-    //INITIALIZE PU BASED ON ID
-    public void Create_PU()
-    {
-        if (PowerUp_created)
-        {
-            Debug.Log("Power up already created");
-            return;
-        }
-        else if (!PlayerTurn)
-        {
-            Debug.Log("It's not your turn");
-        }
-        else{
-        //Agregar for loop para crear los 30 power ups
-        //Agregar los objects a la lista de pile
-        GameObject PU = Instantiate(PUPrefab, PUParentPrefab.transform.position, Quaternion.identity, PUParentPrefab);
-        Debug.Log(pu_Pile[0]);
-        foreach (AtributosPU pu in puCards.powerUps)
-        {
-            Debug.Log("in for loop");
-            if (pu.id == pu_Pile[0])
-            {
-                PU.GetComponent<PUscript>().Init(pu);
-            }
-        }
-        
-        Debug.Log("Creating: " + PU.name);
-        pu_Pile.RemoveAt(0);
-        PowerUp_created = true;
-        }
-
-    }
-
-
-    // create cards in the board
-
-    IEnumerator Create_Board()
-    {
-        int i = 0;
-        //Debug.Log("log"+APIData);
-        allCards=JsonUtility.FromJson<Cards>(APIData);
-        puCards =JsonUtility.FromJson<PUs>(pu_Cards_Data);
-        Debug.Log(allCards.cards);
-        Debug.Log(puCards.powerUps);
-        foreach (Atributos atributosCarta in allCards.cards)
-        {
-            //Debug.Log(atributosCarta);
-            Card_base(Cartas_mano, i, atributosCarta);
-            yield return new WaitForSeconds(0.1f);
-            i++;
-        }
-        /*while (i < Panels.Count)
-        {
-            Card_base(Cartas_mano, i);
-            yield return new WaitForSeconds(0.1f);
-            i++;
-        }*/
-        Init_PU_ID();
-        //We create the power up button
-        PU_button();
-        //We set the player turn to true
-        PlayerTurn = true;
-    }
-    //init the game
-
-    void InitGame()
-    {
-        
-        StartCoroutine(Create_Board());
-       
-    }
-
-     //all pu logic, discard, add, use
-
-    public void AddPU()
-{
-    GameObject PowerUp = PUParentPrefab.transform.GetChild(1).gameObject;
-
-    if (pu_Hand.Count == 3)
-    {
-        Debug.LogError("Can't add more power ups");
-    }
-    else
-    {
-        PowerUp.transform.SetParent(PUSlot);
-        pu_Hand.Add(PowerUp);
-        pu_saved = true;
-
-        // Asignamos el listener para seleccionar el power-up
-        Button powerUpButton = PowerUp.GetComponent<Button>();
-        powerUpButton.onClick.AddListener(() => SelectPowerUp(PowerUp));
-    }
-
-    if (pu_Pile.Count == 0)
-    {
-        Init_PU_ID();
-    }
-}
-
-    public void DiscardPU(){
-        if (PowerUp_created && !pu_saved){
-            GameObject PowerUp = PUParentPrefab.transform.GetChild(1).gameObject;
-            Debug.Log(PowerUp);
-            Destroy(PowerUp);
-            pu_Hand.Remove(PowerUp);
-
-            //AtributosPU atributosPU = PowerUp.GetComponent<AtributosPU>();
-            //pu_Discarded.Add(atributosPU.pu_id);
-        }
-        else{
-            Debug.LogError("No power up to discard");
-        }
-        
-    }
-    public void PU_button()
-    {
-        GameObject PU = Instantiate(PuButtonPrefab, PUParentPrefab.transform.position, Quaternion.identity, PUParentPrefab);
-        Button button = PU.GetComponent<Button>();
-        //We add that we the button is clicked we create a power up
-        button.onClick.AddListener(() => Create_PU());
-        
-    }
-
-    public void Init_PU_ID(){
-        HashSet<int> usedIds = new HashSet<int>();
-      while(pu_Pile.Count < maxPuPile){
-          int id = RandomId();
-          if (!usedIds.Contains(id)){
-              pu_Pile.Add(id);
-              usedIds.Add(id);
-          }
-      }
-
-    }
-
-    public int RandomId(){
-        return Random.Range(8, 36);
-    }
-    public void SelectPowerUp(GameObject powerUpObject)
-{
-    Selectpowerup = powerUpObject;
-    Debug.Log("Selected PowerUp: " + Selectpowerup.name);
-}
-
-    
-    // funcion para usar los power ups
-    
-   public void UsePowerUp(GameObject cardObject, GameObject powerUpObject)
-{
-    
-    Debug.Log("Using power up: " + powerUpObject.name);
-    // We check if the power up is null
-    if (powerUpObject == null)
-    {
-        Debug.Log("No power-up selected");
-        return;
-    }
-    // Accedemos al script de la carta y del power-up y lo guardamos en variables
-    CardScript cardScript = cardObject.GetComponent<CardScript>();
-    PUscript powerUpScript = powerUpObject.GetComponent<PUscript>();
-    // Revisamos si los scripts son nulos esto se uso para debuggear
-    if (cardScript == null)
-    {
-        Debug.LogError("CardScript is null");
-        return;
-    }
-
-    if (powerUpScript == null)
-    {
-        Debug.LogError("PUscript is null");
-        return;
-    }
-    
-    Debug.Log("Power-up effect: " + powerUpScript.atributosPU.ability_effect);
-    // Revisamos el efecto del power-up y aplicamos el efecto correspondiente accediendo a los atributos de los power-ups para ver el efecto
-    switch (powerUpScript.atributosPU.ability_effect)
-    {
-        // En caso de que el efecto sea curación
-        case "curacion":
-            Debug.Log("Applying healing power-up");
-            int healthBoost = powerUpScript.atributosPU.ability_amount;
-            cardScript.atributos.health += healthBoost;
-            Debug.Log("Health boosted by: " + healthBoost);
-            break;
-        // En caso de que el efecto sea restar de energía
-        case  "restaura_energia":
-            Debug.Log("Applying energy restore power-up");
-            int energyBoost = powerUpScript.atributosPU.ability_amount;
-            cardScript.atributos.abilityCost -= energyBoost;
-            Debug.Log("Energy boosted by: " + energyBoost);
-            break;
-        
-        // En caso de que el efecto sea aumentar el daño
-        case "mejora_dano":
-            Debug.Log("Applying damage boost power-up");
-            int damageBoost = powerUpScript.atributosPU.ability_amount;
-            cardScript.atributos.attack += damageBoost;
-            Debug.Log("Damage boosted by: " + damageBoost);
-            break;
-        // En caso de que el efecto sea aumentar la resistencia
-        case "mejora_resistencia":
-            Debug.Log("Applying resistance boost power-up");
-            int resistanceBoost = powerUpScript.atributosPU.ability_amount;
-            cardScript.atributos.resistance += resistanceBoost;
-            Debug.Log("Resistance boosted by: " + resistanceBoost);
-            break;
-
-        // En caso de que el efecto aplicar un escudo
-        case "escudo":
-            Debug.Log("Applying shield power-up");
-            cardScript.atributos.isShielded = powerUpScript.atributosPU.ability_amount;
-            Debug.Log("Shield applied");
-            break;
-        
-         case "bloquea_dano":
-            Debug.Log("Applying damage block power-up");
-            break;
-        
-        case "mejo_velocidad":
-            Debug.Log("Applying speed boost power-up");
-            int speedBoost = powerUpScript.atributosPU.ability_amount;
-            cardScript.atributos.speed += speedBoost;
-            Debug.Log("Speed boosted by: " + speedBoost);
-            break;
-
-            
-
-
-        
-        
-        
-
-        // En caso de que el efecto sea desconocido se imprime un mensaje de error esto tambien se uso para debuggear
-        default:
-            Debug.Log("Unhandled power-up effect: " + powerUpScript.atributosPU.ability_effect);
-            break;
-    }
-    // Se destruye el power-up y se remueve de la lista de power-ups en la mano
-
-    Destroy(powerUpObject);
-    pu_Hand.Remove(powerUpObject);
-
-    // Se setea el power-up seleccionado a null al igual que la carta seleccionada
-    Selectpowerup = null;
-    Selected_card1 = null;
-
-    
-    
-
-    Debug.Log("Power-up applied successfully");
-}
-// end of pu logic
-
-
-
-
 
     //This function changes the position of two cards
 
@@ -417,82 +145,83 @@ public class CardManager : MonoBehaviour
         Debug.Log("index2 change " +Cartas_mano.IndexOf(obj2));
         GameObject temp = obj1;
         Cartas_mano[card_Index1]=obj2;
-        //Debug.Log("Card Index al que se mueve carta 2:" + card_Index1);
         Cartas_mano[card_Index2]=temp;
     }
 
     public void registerCard(GameObject objeto_carta)
-{
-    //We check if the power up is selected
-     if(Selectpowerup != null)
     {
+    //We check if the power up is selected
+        if(Selectpowerup != null)
+        {
         //Llamamos a la función de usar power up con la carta y el power up seleccionado
-        UsePowerUp(objeto_carta, Selectpowerup);  
-        
-    
-    } 
+            GetComponent<PU_controller>().UsePowerUp(objeto_carta, Selectpowerup);  
+        } 
         
     
 
-     else{
-            if (Selected_card1==null){
+        else
+        {
+            if (Selected_card1==null)
+            {
                 if (Cartas_mano.IndexOf(objeto_carta) > 4)
                 {
                     Debug.Log("Enemy card, can't change it");
                 }
             //si no, guardamos la primera carta
-                else{
+                else
+                {
                     Selected_card1= objeto_carta;
                     
                     Debug.Log("Selecting " + card_Index1);
                     //Debug.Log("Selecting " + Selected_card1.name);
-                    if (!Selected_card1.GetComponent<CardScript>().atributos.Alive){
+                    if (!Selected_card1.GetComponent<CardScript>().atributos.Alive)
+                    {
                         Debug.Log("This card is Death");
-                        Selected_card1 = null;}
+                        Selected_card1 = null;
+                    }
                 }
-                }
-            else{
-            //revisamos si la opcion cambio está activada
-                if (Change_Option){
+            }
+            else
+            {
+                //revisamos si la opcion cambio está activada
+                if (Change_Option)
+                {
                     Selected_card2 = objeto_carta;
                     int card_Index1=Cartas_mano.IndexOf(Selected_card1);
                     int card_Index2=Cartas_mano.IndexOf(Selected_card2);
-                        //We check if the cards are the same so you can't change the card with itself
+                    //We check if the cards are the same so you can't change the card with itself
                     Debug.Log("Changing with " + card_Index2);
                     if (Selected_card1 == Selected_card2)
                     {       //Revisamos que la carta no sea la misma
-                            Debug.Log("You can't change the card with itself");
-                            Selected_card1 = null;
-                            Selected_card2 = null;
-                            return;
+                        Debug.Log("You can't change the card with itself");
+                        Selected_card1 = null;
+                        Selected_card2 = null;
+                        return;
                         //We check if the cards are in the hand an in attack position so you can change them and your turn end
-                        }
-                    else if(card_Index1>2 && card_Index2<=2 ||card_Index1<=2 && card_Index2>2 ){
-                            Debug.Log("Caso de Prueba 1");
-                            Debug.Log("index pre" + Cartas_mano.IndexOf(Selected_card1));
-                            Change_Cards(Selected_card1, Selected_card2);
+                    }
+                    else if(card_Index1>2 && card_Index2<=2 ||card_Index1<=2 && card_Index2>2 )
+                    {
+                        Debug.Log("Caso de Prueba 1");
+                        Debug.Log("index pre" + Cartas_mano.IndexOf(Selected_card1));
+                        Change_Cards(Selected_card1, Selected_card2);
                             
-                            //Intercambia las posiciones de los objetos en la lista
-                            Debug.Log("index 1" + card_Index1);
-                            Debug.Log("index 2" + card_Index2);
-                            /*GameObject temp = Selected_card1    ;
-                            Cartas_mano[card_Index1]=Selected_card2;
-                            Debug.Log("Card Index al que se mueve carta 2:" + card_Index1);
-                            Cartas_mano[card_Index2]=temp;*/
-                            Selected_card1 = null;
-                            Selected_card2 = null;
-                            Change_Option = false;
-                            //We end the player turn
-                            
-                            EndTurn();
+                        //Intercambia las posiciones de los objetos en la lista
+                        Debug.Log("index 1" + card_Index1);
+                        Debug.Log("index 2" + card_Index2);
+                        Selected_card1 = null;
+                        Selected_card2 = null;
+                        Change_Option = false;
+                        //We end the player turn  
+                        EndTurn();
 
-                        }
-                    else{
+                    }
+                    else
+                    {
                         Debug.Log("Caso de Prueba 2");
                         //en caso de que las cartas estén en la mano del jugador no le quitara un turno
                         Change_Cards(Selected_card1, Selected_card2);
                         
-                         GameObject temp = Selected_card1;
+                        GameObject temp = Selected_card1;
 
                         //Intercambia las posiciones de los objetos en la lista
                         Cartas_mano[card_Index1]=Selected_card2;
@@ -505,126 +234,132 @@ public class CardManager : MonoBehaviour
                     
                     
                 }
-            else if(Attack_Option){
-                //Check the effect of the card of selected_card1 so we can know his effect in the switch
-                CardScript selectedCardScript = Selected_card1.GetComponent<CardScript>();
-
-                // Revisamos si esta carta ya ha sido usada para atacar previamente en el turno
-                if (Selected_card1.GetComponent<CardScript>().atributos.canAttack==false ){
-                    Debug.Log("You have already used this card for attack");
-                    return;
-                }
-                Debug.Log("Attack option");
-                //Revisamos si la carta es alguna de las que está jugando el jugador
-                if (Selected_card1 == null && (Cartas_mano.IndexOf(objeto_carta) == 3 || Cartas_mano.IndexOf(objeto_carta) == 4)){
-                    Selected_card1 = objeto_carta;
-                    Debug.Log("Selected card for attacking: " + Selected_card1.name);
-                }
-                
-                // switch to check the effect of the card
-                switch (selectedCardScript.atributos.effect) 
+                else if(Attack_Option)
                 {
-               // in case the effect is to deal damage
-                case "dano":
-                Debug.Log("Attack option");
-                // en caso de que sea una de las carta del enemigo la asgnamos a la carta 2
-                 if (Selected_card1 != null && (Cartas_mano.IndexOf(objeto_carta) == 5 || Cartas_mano.IndexOf(objeto_carta) == 6)){
-            
-                    Selected_card2= objeto_carta;
-                    Debug.Log("Selected card to attack: " + Selected_card2.name);
-                    //hacer el ataque de las cartas
-                    Attack(Selected_card1, Selected_card2);
-                    // cambiamos la opcion Can Attack para que ya no se pueda atacar con esa carta
-                    
-                    Selected_card1 = null;
-                    Selected_card2 = null;
-                    Attack_Option = false;
-                    //En caso de que ya haya usado los dos ataques ya no permite atacar
-                    
-                    if (counter == 2){
-                        counter = 0;
-                        Attack_Option=false;
-                        Debug.Log("No more attacks available");
-                    }
-                    if (Cartas_mano[5].GetComponent<CardScript>().atributos.health <= 0 && Cartas_mano[6].GetComponent<CardScript>().atributos.health <= 0)
-                    {
-                        Debug.Log("You win");
-                        PlayerTurn = false;
+                    //Check the effect of the card of selected_card1 so we can know his effect in the switch
+                    CardScript selectedCardScript = Selected_card1.GetComponent<CardScript>();
 
+                    // Revisamos si esta carta ya ha sido usada para atacar previamente en el turno
+                    if (Selected_card1.GetComponent<CardScript>().atributos.canAttack==false )
+                    {
+                        Debug.Log("You have already used this card for attack");
+                        return;
+                    }
+
+                    Debug.Log("Attack option");
+                    //Revisamos si la carta es alguna de las que está jugando el jugador
+                    if (Selected_card1 == null && (Cartas_mano.IndexOf(objeto_carta) == 3 || Cartas_mano.IndexOf(objeto_carta) == 4))
+                    {
+                        Selected_card1 = objeto_carta;
+                        Debug.Log("Selected card for attacking: " + Selected_card1.name);
+                    }
+                    
+                    // switch to check the effect of the card
+                    switch (selectedCardScript.atributos.effect) 
+                    {
+                        // in case the effect is to deal damage
+                        case "dano":
+                            Debug.Log("Attack option");
+                            // en caso de que sea una de las carta del enemigo la asgnamos a la carta 2
+                            if (Selected_card1 != null && (Cartas_mano.IndexOf(objeto_carta) == 5 || Cartas_mano.IndexOf(objeto_carta) == 6))
+                            {
+                    
+                                Selected_card2= objeto_carta;
+                                Debug.Log("Selected card to attack: " + Selected_card2.name);
+                                //hacer el ataque de las cartas
+                                Attack(Selected_card1, Selected_card2);
+                                // cambiamos la opcion Can Attack para que ya no se pueda atacar con esa carta
+                                
+                                Selected_card1 = null;
+                                Selected_card2 = null;
+                                Attack_Option = false;
+                                //En caso de que ya haya usado los dos ataques ya no permite atacar
+                                
+                                if (counter == 2){
+                                    counter = 0;
+                                    Attack_Option=false;
+                                    Debug.Log("No more attacks available");
+                                }
+                                if (Cartas_mano[5].GetComponent<CardScript>().atributos.health <= 0 && Cartas_mano[6].GetComponent<CardScript>().atributos.health <= 0)
+                                {
+                                Debug.Log("You win");
+                                PlayerTurn = false;
+                                }
+                            }
+                            break;
+                        //In case the effect is to heal a card
+                        case "curacion":
+                            Debug.Log("Heal option");
+                            if (Selected_card1 != null && (Cartas_mano.IndexOf(objeto_carta) == 3 || Cartas_mano.IndexOf(objeto_carta) == 4))
+                            {
+                        
+                                Selected_card2= objeto_carta;
+                                Debug.Log("Selected card to heal: " + Selected_card2.name);
+                                //hacer el ataque de las cartas
+                                Heal(Selected_card1, Selected_card2);
+                                // cambiamos la opcion Can Attack para que ya no se pueda atacar con esa carta
+                                
+                                Selected_card1 = null;
+                                Selected_card2 = null;
+                                Attack_Option = false;
+                                //En caso de que ya haya usado los dos ataques ya no permite atacar
+                                
+                                if (counter == 2)
+                                {
+                                    counter = 0;
+                                    Attack_Option=false;
+                                    Debug.Log("No more attacks available");
+                                }
+                                if (Cartas_mano[3].GetComponent<CardScript>().atributos.health <= 0 && Cartas_mano[4].GetComponent<CardScript>().atributos.health <= 0)
+                                {
+                                    Debug.Log("You win");
+                                    PlayerTurn = false;
+
+                                }
+                            }
+                            break;
+                        //In case the effect is to boost the damage of a card
+                        case "mejora_dano":
+                            Debug.Log("Damage boost option");
+                            if (Selected_card1 != null && (Cartas_mano.IndexOf(objeto_carta) == 3 || Cartas_mano.IndexOf(objeto_carta) == 4))
+                            {
+
+                        
+                                Selected_card2= objeto_carta;
+                                Debug.Log("Selected card to boost damage: " + Selected_card2.name);
+                                //hacer el ataque de las cartas
+                                mejoradano(Selected_card1, Selected_card2);
+                                // cambiamos la opcion Can Attack para que ya no se pueda atacar con esa carta
+                                
+                                Selected_card1 = null;
+                                Selected_card2 = null;
+                                Attack_Option = false;
+                                //En caso de que ya haya usado los dos ataques ya no permite atacar
+                                
+                                if (counter == 2)
+                                {
+                                    counter = 0;
+                                    Attack_Option=false;
+                                    Debug.Log("No more attacks available");
+                                }
+                                if (Cartas_mano[3].GetComponent<CardScript>().atributos.health <= 0 && Cartas_mano[4].GetComponent<CardScript>().atributos.health <= 0)
+                                {
+                                    Debug.Log("You win");
+                                    PlayerTurn = false;
+                                }
+                            }
+                            break;
+
+                        default:
+                            Debug.Log("No se a implementado la funcion todavia");
+                            break;
+                    
                     }
                 }
-                break;
-                //In case the effect is to heal a card
-                case "curacion":
-                Debug.Log("Heal option");
-                 if (Selected_card1 != null && (Cartas_mano.IndexOf(objeto_carta) == 3 || Cartas_mano.IndexOf(objeto_carta) == 4)){
-            
-                    Selected_card2= objeto_carta;
-                    Debug.Log("Selected card to heal: " + Selected_card2.name);
-                    //hacer el ataque de las cartas
-                    Heal(Selected_card1, Selected_card2);
-                    // cambiamos la opcion Can Attack para que ya no se pueda atacar con esa carta
-                    
-                    Selected_card1 = null;
-                    Selected_card2 = null;
-                    Attack_Option = false;
-                    //En caso de que ya haya usado los dos ataques ya no permite atacar
-                    
-                    if (counter == 2){
-                        counter = 0;
-                        Attack_Option=false;
-                        Debug.Log("No more attacks available");
-                    }
-                    if (Cartas_mano[3].GetComponent<CardScript>().atributos.health <= 0 && Cartas_mano[4].GetComponent<CardScript>().atributos.health <= 0)
-                    {
-                        Debug.Log("You win");
-                        PlayerTurn = false;
-
-                    }
-                }
-                break;
-                //In case the effect is to boost the damage of a card
-                case "mejora_dano":
-                Debug.Log("Damage boost option");
-                 if (Selected_card1 != null && (Cartas_mano.IndexOf(objeto_carta) == 3 || Cartas_mano.IndexOf(objeto_carta) == 4)){
-
-            
-                    Selected_card2= objeto_carta;
-                    Debug.Log("Selected card to boost damage: " + Selected_card2.name);
-                    //hacer el ataque de las cartas
-                    mejoradano(Selected_card1, Selected_card2);
-                    // cambiamos la opcion Can Attack para que ya no se pueda atacar con esa carta
-                    
-                    Selected_card1 = null;
-                    Selected_card2 = null;
-                    Attack_Option = false;
-                    //En caso de que ya haya usado los dos ataques ya no permite atacar
-                    
-                    if (counter == 2){
-                        counter = 0;
-                        Attack_Option=false;
-                        Debug.Log("No more attacks available");
-                    }
-                    if (Cartas_mano[3].GetComponent<CardScript>().atributos.health <= 0 && Cartas_mano[4].GetComponent<CardScript>().atributos.health <= 0)
-                    {
-                        Debug.Log("You win");
-                        PlayerTurn = false;
-
-                    }
-                }
-                break;
-
-                default:
-                Debug.Log("No se a implementado la funcion todavia");
-                break;
-                
             }
-            
         }
     }
-     }
-}
-      //If the change option is not active we send error message
+    //If the change option is not active we send error message
 
         
     //This function is used to change the state of the change option(active in the change button)
@@ -635,262 +370,240 @@ public class CardManager : MonoBehaviour
         {
             Debug.Log("You have Attack option active, you can't change now");
             
-            }
+        }
         else if(PlayerTurn == false)
         {
             Debug.Log("It's not your turn");
             return;
         
-        }else if(counter != 0)
+        }
+        else if(counter != 0)
         {
             Debug.Log("You have already attacked");
             return;
-        }else{
-        if (!changeButtonPressed)
-        {
-            Change_Option = true;
-            changeButtonPressed = true;
         }
-        else
-        {
-            Change_Option = false;
-            changeButtonPressed = false;
-        }
-            
-
+        else{
+            if (!changeButtonPressed)
+            {
+                Change_Option = true;
+                changeButtonPressed = true;
+            }
+            else
+            {
+                Change_Option = false;
+                changeButtonPressed = false;
+            }
         }
     }
 
-    public void Turn(){
-        /*
-        is called when the player turn is true
-
-        1. Add energy to the bar
-        2. Use button listener to either:
-            a) Use a card's ability
-            b) Change a card
-            c) add the new power up to the hand
-        3. Increase number of turn
-        4. End Turn (set playerTurn to false)
-        5. Call CountCountEnemyTurn
-
-        */
-    }
+   
 
     //Increase Energy amount
     public void IncreaseEnegry(){
         Debug.Log("Energy increased");
-        if(energy + (num_turn-1) * 10 <= max_energy){
+        if(energy + (num_turn-1) * 10 <= max_energy)
+        {
             energy += (num_turn-1) * 10;
         }
-        else if(energy + (num_turn-1) * 10 > max_energy){
+        else if(energy + (num_turn-1) * 10 > max_energy)
+        {
             energy = 100;
         }
 
         Slider sliderComponent = energySlider.GetComponent<Slider>();
         sliderComponent.value = energy;
-        
-
-        
+           
     }
 
     
-//-------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------
+    //-------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------
 
-// Function to attack enemy cards
-public void Attack(GameObject objeto_carta1, GameObject objeto_carta2)
-{
-    // First we check if the attack option is active
-    if (Attack_Option)
+    // Function to attack enemy cards
+    public void Attack(GameObject objeto_carta1, GameObject objeto_carta2)
     {
-        // We check if the cards are in the player hand
-        if (objeto_carta1 != null && objeto_carta2 != null)
+        // First we check if the attack option is active
+        if (Attack_Option)
         {
-            // We set the attributes of the cards atrributesCarta1 for the player card and atributosCarta2 for the enemy card
-            Atributos atributosCarta1 = objeto_carta1.GetComponent<CardScript>().atributos;
-            Atributos atributosCarta2 = objeto_carta2.GetComponent<CardScript>().atributos;
-            
-            if (atributosCarta1.abilityCost<=energy)
+            // We check if the cards are in the player hand
+            if (objeto_carta1 != null && objeto_carta2 != null)
             {
+                // We set the attributes of the cards atrributesCarta1 for the player card and atributosCarta2 for the enemy card
+                Atributos atributosCarta1 = objeto_carta1.GetComponent<CardScript>().atributos;
+                Atributos atributosCarta2 = objeto_carta2.GetComponent<CardScript>().atributos;
                 
-                Debug.Log("objeto 1:"+Cartas_mano.IndexOf(objeto_carta1));
-                Debug.Log("objeto 2:"+Cartas_mano.IndexOf(objeto_carta2));
-                // We check if the cards are in the right position 
-                if ((Cartas_mano.IndexOf(objeto_carta1) == 3 || Cartas_mano.IndexOf(objeto_carta1) == 4) &&
-                    (Cartas_mano.IndexOf(objeto_carta2) == 5 || Cartas_mano.IndexOf(objeto_carta2) == 6))
+                if (atributosCarta1.abilityCost<=energy)
                 {
-                 int damageDealt = atributosCarta1.attack - atributosCarta2.resistance;
-                    // We check if the damage dealt is greater than 0
-                    if (damageDealt <= 0)
-                    {
-                        Debug.Log("No damage dealt because the enemy card has more resistance than the player card's attack");
-                        counter++;
-                    }else{
-
-                   // We attack the enemy card 
-                    atributosCarta2.health -= damageDealt;
-                    Debug.Log(objeto_carta1.name + " attacked " + objeto_carta2.name + " for " + atributosCarta1.attack + " damage.");
                     
-                    //Decrease energy amount-- fix later on with the corresponding value
-                    energy-= atributosCarta1.abilityCost;
-                    energyText.text=$"{energy}";
-                    Slider sliderComponent = energySlider.GetComponent<Slider>();
-                    sliderComponent.value = energy;
-                    objeto_carta1.GetComponent<CardScript>().atributos.canAttack=false;
-                    counter++;
+                    Debug.Log("objeto 1:"+Cartas_mano.IndexOf(objeto_carta1));
+                    Debug.Log("objeto 2:"+Cartas_mano.IndexOf(objeto_carta2));
+                    // We check if the cards are in the right position 
+                    if ((Cartas_mano.IndexOf(objeto_carta1) == 3 || Cartas_mano.IndexOf(objeto_carta1) == 4) &&
+                        (Cartas_mano.IndexOf(objeto_carta2) == 5 || Cartas_mano.IndexOf(objeto_carta2) == 6))
+                    {
+                        int damageDealt = atributosCarta1.attack - atributosCarta2.resistance;
+                        // We check if the damage dealt is greater than 0
+                        if (damageDealt <= 0)
+                        {
+                            Debug.Log("No damage dealt because the enemy card has more resistance than the player card's attack");
+                            counter++;
+                        }
+                        else
+                        {
+
+                            // We attack the enemy card 
+                            atributosCarta2.health -= damageDealt;
+                            Debug.Log(objeto_carta1.name + " attacked " + objeto_carta2.name + " for " + atributosCarta1.attack + " damage.");
+                            
+                            //Decrease energy amount-- fix later on with the corresponding value
+                            energy-= atributosCarta1.abilityCost;
+                            energyText.text=$"{energy}";
+                            Slider sliderComponent = energySlider.GetComponent<Slider>();
+                            sliderComponent.value = energy;
+                            objeto_carta1.GetComponent<CardScript>().atributos.canAttack=false;
+                            counter++;
+                        }
+                    }
+                    else
+                    {
+                        Debug.Log("No valid cards to attack");
                     }
                 }
                 else
                 {
-                    Debug.Log("No valid cards to attack");
+                    Debug.Log("No enough energy");
                 }
+                
             }
-            else{
-                Debug.Log("No enough energy");
+            else
+            {
+                Debug.Log("Select two cards to attack");
             }
+        }
+        else
+        {
+            Debug.Log("Invalid option");
+        }
+        
+    }
+    //Function to heal player cards 
+    public void Heal (GameObject objeto_carta1, GameObject objeto_carta2)
+    {
+ 
+        Atributos atributosCarta1 = objeto_carta1.GetComponent<CardScript>().atributos;
+        Atributos atributosCarta2 = objeto_carta2.GetComponent<CardScript>().atributos;
+    
+        if (atributosCarta1.abilityCost<=energy)
+        {
+            if ((Cartas_mano.IndexOf(objeto_carta1) == 3 || Cartas_mano.IndexOf(objeto_carta1) == 4) &&
+                (Cartas_mano.IndexOf(objeto_carta2) == 3 || Cartas_mano.IndexOf(objeto_carta2) == 4))
+            {
+                int healthBoost = atributosCarta1.attack;
+                atributosCarta2.health += healthBoost;
+                if (atributosCarta2.health > 100) 
+                {
+                    atributosCarta2.health = 100;
+                }
+                Debug.Log(objeto_carta1.name + " healed " + objeto_carta2.name + " for " + atributosCarta1.attack + " health.");
+                energy-= atributosCarta1.abilityCost;
+                energyText.text=$"{energy}";
+                Slider sliderComponent = energySlider.GetComponent<Slider>();
+                sliderComponent.value = energy;
+                objeto_carta1.GetComponent<CardScript>().atributos.canAttack=false;
+                counter++;
+            }
+            else
+            {
+                Debug.Log("No valid cards to heal");
+            }
+        }
+        else
+        {
+            Debug.Log("No enough energy");
+        }
+
+    }
+//Function to boost the damage of a player card
+    public void mejoradano(GameObject objeto_carta1, GameObject objeto_carta2)
+    {
+        Atributos atributosCarta1 = objeto_carta1.GetComponent<CardScript>().atributos;
+        Atributos atributosCarta2 = objeto_carta2.GetComponent<CardScript>().atributos;
+
+        if (atributosCarta1.abilityCost<=energy)
+        {
+            if ((Cartas_mano.IndexOf(objeto_carta1) == 3 || Cartas_mano.IndexOf(objeto_carta1) == 4) &&
+            (Cartas_mano.IndexOf(objeto_carta2) == 3 || Cartas_mano.IndexOf(objeto_carta2) == 4))
+            {
+                int damageBoost = atributosCarta1.attack;
+                atributosCarta2.attack += damageBoost;
+                Debug.Log(objeto_carta1.name + " boosted " + objeto_carta2.name + " for " + atributosCarta1.attack + " damage.");
+                energy-= atributosCarta1.abilityCost;
+                energyText.text=$"{energy}";
+                Slider sliderComponent = energySlider.GetComponent<Slider>();
+                sliderComponent.value = energy;
+                objeto_carta1.GetComponent<CardScript>().atributos.canAttack=false;
+                counter++;
+            }
+        }
+    }
+
+
+    // Function to change the state of the attack option (active in the attack button)
+    public void Attack_button()
+    {
+        if(Change_Option == true)
+        {
+            Debug.Log("You have Change option active, you can't attack now");
             
         }
-        else
+        else if (PlayerTurn == false)
         {
-            Debug.Log("Select two cards to attack");
-        }
-    }
-    else
-    {
-        Debug.Log("Invalid option");
-    }
-    
-}
-//Function to heal player cards 
-public void Heal (GameObject objeto_carta1, GameObject objeto_carta2){
- 
-    Atributos atributosCarta1 = objeto_carta1.GetComponent<CardScript>().atributos;
-    Atributos atributosCarta2 = objeto_carta2.GetComponent<CardScript>().atributos;
-   
-    if (atributosCarta1.abilityCost<=energy)
-    {
-        if ((Cartas_mano.IndexOf(objeto_carta1) == 3 || Cartas_mano.IndexOf(objeto_carta1) == 4) &&
-            (Cartas_mano.IndexOf(objeto_carta2) == 3 || Cartas_mano.IndexOf(objeto_carta2) == 4))
-        {
-            int healthBoost = atributosCarta1.attack;
-            atributosCarta2.health += healthBoost;
-              if (atributosCarta2.health > 100) {
-                atributosCarta2.health = 100;
-              }
-            Debug.Log(objeto_carta1.name + " healed " + objeto_carta2.name + " for " + atributosCarta1.attack + " health.");
-            energy-= atributosCarta1.abilityCost;
-            energyText.text=$"{energy}";
-            Slider sliderComponent = energySlider.GetComponent<Slider>();
-            sliderComponent.value = energy;
-            objeto_carta1.GetComponent<CardScript>().atributos.canAttack=false;
-            counter++;
+                Debug.Log("It's not your turn");
+                return;
         }
         else
         {
-            Debug.Log("No valid cards to heal");
+            Attack_Option=!Attack_Option;
         }
     }
-    else
-    {
-        Debug.Log("No enough energy");
-    }
-
-}
-//Function to boost the damage of a player card
-public void mejoradano(GameObject objeto_carta1, GameObject objeto_carta2)
-{
-    Atributos atributosCarta1 = objeto_carta1.GetComponent<CardScript>().atributos;
-    Atributos atributosCarta2 = objeto_carta2.GetComponent<CardScript>().atributos;
-
-    if (atributosCarta1.abilityCost<=energy)
-    {
-        if ((Cartas_mano.IndexOf(objeto_carta1) == 3 || Cartas_mano.IndexOf(objeto_carta1) == 4) &&
-           (Cartas_mano.IndexOf(objeto_carta2) == 3 || Cartas_mano.IndexOf(objeto_carta2) == 4))
-        {
-            int damageBoost = atributosCarta1.attack;
-            atributosCarta2.attack += damageBoost;
-            Debug.Log(objeto_carta1.name + " boosted " + objeto_carta2.name + " for " + atributosCarta1.attack + " damage.");
-            energy-= atributosCarta1.abilityCost;
-            energyText.text=$"{energy}";
-            Slider sliderComponent = energySlider.GetComponent<Slider>();
-            sliderComponent.value = energy;
-            objeto_carta1.GetComponent<CardScript>().atributos.canAttack=false;
-            counter++;
-        }
-     
-    }
-}
-
-
-// Function to change the state of the attack option (active in the attack button)
- public void Attack_button(){
-    if(Change_Option == true)
-    {
-        Debug.Log("You have Change option active, you can't attack now");
-        
-        }
-   else if (PlayerTurn == false)
-        
-        {
-            Debug.Log("It's not your turn");
-            return;}
-
-        else{
-        Attack_Option=!Attack_Option;
-    }
- }
 
 // Function to create the power up button
 
 
 
 
-public void EndTurn()
-{
-    PlayerTurn = false;
-    ++num_turn;
-    Atributos activeCard1 = Cartas_mano[3].GetComponent<CardScript>().atributos;
-    Atributos activeCard2 = Cartas_mano[4].GetComponent<CardScript>().atributos;
-    
-    DiscardPU();
+    public void EndTurn()
+    {
+        PlayerTurn = false;
+        ++num_turn;
+        Atributos activeCard1 = Cartas_mano[3].GetComponent<CardScript>().atributos;
+        Atributos activeCard2 = Cartas_mano[4].GetComponent<CardScript>().atributos;
+        
+        GetComponent<PU_controller>().DiscardPU();
 
-    activeCard1.canAttack = true;
-    activeCard2.canAttack = true;
-    activeCard1.Update_Shield();
-    activeCard2.Update_Shield();
-
-
-
-    IncreaseEnegry();
-    turnText.text = $"Turn: {num_turn}";
-    energyText.text = $"{energy}";
-    CountCountEnemyTurn = true;
-    counter = 0;
-    PowerUp_created = false;
-
-    
-    Selected_card1 = null;
-    Selected_card2 = null;
-    Change_Option = false;
-    pu_saved = false;
-    
-    EnemyScript enemyScript = GetComponent<EnemyScript>();
-    enemyScript.EnemyTurn();
+        activeCard1.canAttack = true;
+        activeCard2.canAttack = true;
+        activeCard1.Update_Shield();
+        activeCard2.Update_Shield();
 
 
-       
-    
-    
-    
-    
-}
+
+        IncreaseEnegry();
+        turnText.text = $"Turn: {num_turn}";
+        energyText.text = $"{energy}";
+        CountCountEnemyTurn = true;
+        counter = 0;
+        GetComponent<PU_controller>().PowerUp_created = false;
+
+        
+        Selected_card1 = null;
+        Selected_card2 = null;
+        Change_Option = false;
+        GetComponent<PU_controller>().pu_saved = false;
+        
+        EnemyScript enemyScript = GetComponent<EnemyScript>();
+        enemyScript.EnemyTurn();
+    }
 
     //-------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------
-    //enemy turn
     
-
-    
-
-
-
 }
