+++ VideogameProject/Unity_FA/Assets/Scripts/PU_controller.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 57f44594de7da884cae311c098b70a5d
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
